Moved semaphore classes into own files.
